% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{clean_wq}
\alias{clean_wq}
\title{Clean water quality data from common issues}
\usage{
clean_wq(df, zero_to_min = 0.5)
}
\arguments{
\item{df}{A dataframe with water quality data, including columns "Station",
"DATE", "Variables", "Values", "Source".}

\item{zero_to_min}{(optional) A numeric coefficient to replace zeros by the
minimum variable value multiplied by zero_to_min.
Default is \code{zero_to_min = 0.5} .}
}
\value{
A cleaned dataframe.
}
\description{
This function cleans water quality data by addressing common issues.
}
\examples{
\dontrun{
  # Load calibration data from an Excel file
  temp_path <- system.file("extdata", "calibration_data.xlsx", package = "SWATprepR")
  cal_data <- load_template(temp_path)
  
  # Clean water quality data
  cal_data$data <- clean_wq(cal_data$data)
}
}
\seealso{
\code{\link{load_template}}
}
\keyword{cleaning}
