% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{df_to_txt}
\alias{df_to_txt}
\title{Transform dataframe into text and writes into existing file}
\usage{
df_to_txt(write_path, fname, df, spacing)
}
\arguments{
\item{write_path}{character, path to SWAT+ txtinout folder (example "my_model").}

\item{fname}{character, file name.}

\item{df}{dataframe}

\item{spacing}{character, example spacing <- c('\%-30s', rep('\%-13s', 2), '\%-15s', '\%-3s')}
}
\value{
updated text file
}
\description{
Transform dataframe into text and writes into existing file
}
\examples{
\dontrun{
spacing <- c(rep('\%10s', 5))
df_to_txt("output", "time.sim", df, spacing)
}
}
\keyword{internal}
