% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparing.R
\name{fill_with_closest}
\alias{fill_with_closest}
\title{Fill missing variables from the closest stations with available data}
\usage{
fill_with_closest(
  meteo_lst,
  par_fill = c("TMP_MAX", "TMP_MIN", "PCP", "RELHUM", "WNDSPD", "SLR")
)
}
\arguments{
\item{meteo_lst}{A nested list of lists with dataframes.
Nested structure: \code{meteo_lst -> data -> Station ID -> Parameter -> 
  Dataframe (DATE, PARAMETER)}.
Nested \code{meteo_lst -> stations -> Dataframe (ID, Name, Elevation, Source, 
  geometry, Long, Lat)}.}

\item{par_fill}{(optional) A vector of variables to be filled. Default is
\code{par_fill = c("TMP_MAX", "TMP_MIN","PCP", "RELHUM", "WNDSPD", "SLR")}.}
}
\value{
A list of dataframes with filled data. Updated list is for meteo_lst$data.
}
\description{
This function fills missing variables by interpolating values from the
closest stations that have data.
}
\examples{
\dontrun{
  # Load weather data from an Excel file
  temp_path <- system.file("extdata", "weather_data.xlsx", package = "SWATprepR")
  met_lst <- load_template(temp_path, 3035)
  
  # Fill for missing variables
  met_lst$data <- fill_with_closest(met_lst, c("TMP_MAX", "TMP_MIN"))
}
}
\keyword{internal}
