% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{find_closest}
\alias{find_closest}
\title{Find Closest Stations for Meteorological Parameters}
\usage{
find_closest(
  meteo_lst,
  par_fill = c("TMP_MAX", "TMP_MIN", "PCP", "RELHUM", "WNDSPD", "SLR")
)
}
\arguments{
\item{meteo_lst}{A nested list with dataframes.
Nested structure: \code{meteo_lst -> data -> Station ID -> Parameter -> 
  Dataframe (DATE, PARAMETER)}.
Nested \code{meteo_lst -> stations -> Dataframe (ID, Name, Elevation, Source, 
  geometry, Long, Lat)}. \cr\cr
meteo_lst can be created using \code{\link{load_template}} function using
'xlsx' template file or it could to be created with \code{\link{load_swat_weather}}
function loading information from SWAT+ model setup weather files.}

\item{par_fill}{(optional) A vector of variables to be filled. Default is
\code{par_fill = c("TMP_MAX", "TMP_MIN","PCP", "RELHUM", "WNDSPD", "SLR")}.}
}
\value{
A data frame with station information and filled meteorological parameters.
}
\description{
This function finds the closest stations for specified meteorological parameters.
}
\examples{
\dontrun{
  # Load weather data from an Excel file
  temp_path <- system.file("extdata", "weather_data.xlsx", package = "SWATprepR")
  met_lst <- load_template(temp_path, 3035)
  
  # Fill for missing variables
  df <- find_closest(met_lst)
}
}
\keyword{internal}
