% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{find_dif}
\alias{find_dif}
\title{Find Differences Between Files in Two Setups}
\usage{
find_dif(
  path1,
  path2,
  name1 = "Setup 1",
  name2 = "Setup2",
  remove_pattern = NULL
)
}
\arguments{
\item{path1}{Path to the first setup.}

\item{path2}{Path to the second setup.}

\item{name1}{Character for name for the first setup (default is \code{name1 = Setup 1}).}

\item{name2}{Character for name for the second setup (default is \code{name2 = Setup 2}).}

\item{remove_pattern}{Regex pattern for files to be removed. Default is \code{remove_pattern = NULL},
which means pattern removes file types: txt, zip, fin, out, exe, bak, mgts,
farm, sqlite, swf, sch, hmd, wnd, pcp, slr, tmp. Example for excluding only txt files  \code{remove_pattern <- '.*.txt'}.
Other can be added using | operator.}
}
\value{
A list containing differences between files if found, otherwise a message indicating no differences.
}
\description{
This function scans two folders and identifies files that exist only in one
setup or files with differences between setups.
}
