% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{find_time_step}
\alias{find_time_step}
\title{Identify time step}
\usage{
find_time_step(d_after, d_before)
}
\arguments{
\item{d_after}{single value data frame with later date}

\item{d_before}{single value data frame with earlier date}
}
\value{
list with 'typ' code and 'rec_typ' code
}
\description{
Identify time step
}
\examples{
\dontrun{
find_time_step(r[2, "DATE"], r[1, "DATE"])
}
}
\keyword{internal}
