% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparing.R
\name{get_data_to_interpolate}
\alias{get_data_to_interpolate}
\title{Function to get data to be used in the interpolation}
\usage{
get_data_to_interpolate(meteo_lst, par)
}
\arguments{
\item{meteo_lst}{nested list of lists with dataframes.
Nested structure meteo_lst -> data -> Station ID -> Parameter -> Dataframe (DATE, PARAMETER).}

\item{par}{is weather variable to extract (i.e. "PCP", "SLR", etc)}
}
\value{
SpatialPointsDataFrame with columns as days and rows as stations.
}
\description{
Function to get data to be used in the interpolation
}
\examples{
##get_data_to_interpolate (meteo_lst, "PCP")
}
\keyword{internal}
