% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{get_dates}
\alias{get_dates}
\title{Function to get maximum min and max dates available for all time series data for all stations for all parameters.}
\usage{
get_dates(meteo_lst)
}
\arguments{
\item{meteo_lst}{nested list of lists with dataframes.
Nested structure meteo_lst -> data -> Station ID -> Parameter -> Dataframe (DATE, PARAMETER).}
}
\value{
list with 2 values: minimum and maximum value for all available data.
}
\description{
Function to get maximum min and max dates available for all time series data for all stations for all parameters.
}
\examples{
##get_nx_date_full(meteo_lst)
}
\keyword{internal}
