% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{get_grid}
\alias{get_grid}
\title{Prepare grid for interpolation}
\usage{
get_grid(b, grid_spacing, wkt_str)
}
\arguments{
\item{b}{vector of border c(xmin, ymin, xmax, ymax) coordinates to define extent of the grid.}

\item{grid_spacing}{Grid spacing. Depending on coordinate system could be in meters, or degrees.}

\item{wkt_str}{projection string}
}
\value{
Grid needed for the interpolation.
}
\description{
Prepare grid for interpolation
}
\examples{
\dontrun{
get_grid(shp 2000)
}
}
\keyword{internal}
