% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparing.R
\name{get_hsg}
\alias{get_hsg}
\title{Function to get soil hydrologic groups}
\usage{
get_hsg(d_imp, d_wtr, drn, t)
}
\arguments{
\item{d_imp}{character for depth to impervious layer. Only three entry options possible: "<50cm", "50-100cm", ">100cm".}

\item{d_wtr}{character for water table high. Only three entry options possible: "<60cm", "60-100cm", ">100cm".}

\item{drn}{character for tile drainage status. Only two entry options possible: "Y" for drained areas, "N" for areas without working tile drains.}

\item{t}{one row dataframe with all SOL_Z and SOL_K values for soil type.}
}
\value{
character of soil hydrologic group A, B, C or D
}
\description{
Function to get soil hydrologic groups
}
\examples{
df <- data.frame(SOL_K1 = 10, SOL_K2 = 1, SOL_K3 = 2, 
                 SOL_Z1 = 250, SOL_Z2 = 700, SOL_Z3 = 1000)
get_hsg(d_imp = ">100cm", d_wtr = "<60cm", drn = "Y", df)
}
