% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converting.R
\name{get_interpolated_data}
\alias{get_interpolated_data}
\title{Main interpolation function}
\usage{
get_interpolated_data(meteo_lst, par, grid_spacing, dem_data, idw_exponent = 2)
}
\arguments{
\item{meteo_lst}{nested list of lists with dataframes.
Nested structure meteo_lst -> data -> Station ID -> Parameter -> Dataframe (DATE, PARAMETER).}

\item{par}{character representing weather variable to extract (i.e. "PCP", "SLR", etc).}

\item{grid_spacing}{numeric value for distance in grid. Units of coordinate system should be used.}

\item{dem_data}{DEM raster data in same projection as weather station.}

\item{idw_exponent}{numeric value for exponent parameter to be used in interpolation
(optional, default value is 2).}
}
\value{
SpatialPointsDataFrame with interpolated data.
}
\description{
Main interpolation function
}
\examples{
##get_interpolated_data(meteo_lst, "PCP", 2000, DEM, 2)
}
