% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helping.R
\name{get_interpolation}
\alias{get_interpolation}
\title{Get interpolation values for catchment}
\usage{
get_interpolation(sp_df, st, i, idw_exponent)
}
\arguments{
\item{sp_df}{SpatialPointsDataFrame with one column for one day and rows as stations.}

\item{st}{SpatialPointsDataFrame with virtual stations from a grid. Should contain
dataframe in data for station elevation from DEM}

\item{i}{integer representing day number.}
}
\value{
SpatialPointsDataFrame with virtual stations and interpolated data in data dataframe
}
\description{
Get interpolation values for catchment
}
\examples{
##get_interpolation(sp_df, meteo_pts, 2, 2)
}
\keyword{internal}
