% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparing.R
\name{get_usersoil_table}
\alias{get_usersoil_table}
\title{Function to prepare user soil table for SWAT model}
\usage{
get_usersoil_table(csv_path, hsg = FALSE, keep_values = FALSE, nb_lyr = NA)
}
\arguments{
\item{csv_path}{character path to csv file (example "usersoil_lrew.csv"). File should be comma separated
with minimum columns of SNAM,    NLAYERS    and SOL_Z, CLAY, SILT, SAND, SOL_CBN for each available
soil layer (minimum 1).}

\item{hsg}{Boolean, TRUE - prepare soil hydrological groups, FALSE - no soil hydrological group
preparation will be done. Optional (\code{Default = FALSE}). If \code{hsg = TRUE} three additional columns
should be in an input table: Impervious (allowed values are "<50cm", "50-100cm", ">100cm"), Depth (allowed
values are "<60cm", "60-100cm", ">100cm") and Drained (allowed values are "Y" for drained areas,
"N" for areas without working tile drains). More information can be found in the SWAT+ modeling protocol
\href{https://doi.org/10.5281/zenodo.7463395}{Table 3.3}.}

\item{keep_values}{Boolean, TRUE - keep old values (new values only will be left where 0
or NA values are present in an input table), FALSE - keep only new values. Optional (\code{Default = FALSE}).}

\item{nb_lyr}{integer, number of layers resulting user soil data should contain. Optional (\code{Default = NA},
which stands for the same number as in input data).}
}
\value{
dataframe with fully formatted and filled table of soil parameters for SWAT model.
}
\description{
Function to prepare user soil table for SWAT model
}
\examples{
\dontrun{
usersoils <- get_usersoil_table("table.csv")
write.csv(usersoils, "usersoils.csv", row.names=FALSE, quote=FALSE)
}
}
