% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{hide_show}
\alias{hide_show}
\title{Function to put option remove hide or show all lines in chart and this function also print chart.}
\usage{
hide_show(graph)
}
\arguments{
\item{graph}{plotly graph object}
}
\value{
plotly graph object with option to remove or show all lines
}
\description{
Function to put option remove hide or show all lines in chart and this function also print chart.
}
\examples{
\dontrun{
hide_show(fig)
}
}
\keyword{internal}
