% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_helper.R
\name{hide_show_print_graph}
\alias{hide_show_print_graph}
\title{Function to put option remove hide or show all lines in chart and this function also print chart.}
\usage{
hide_show_print_graph(graph)
}
\arguments{
\item{graph}{plotly graph object}
}
\value{
plotly graph object with option to romove or show all lines
}
\description{
Function to put option remove hide or show all lines in chart and this function also print chart.
}
\examples{
##hide_show_print_graph(fig)
}
\keyword{internal}
