% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparing.R
\name{interpolate}
\alias{interpolate}
\title{Interpolate Weather Data}
\usage{
interpolate(
  meteo_lst,
  catchment_boundary_path,
  dem_data_path,
  grid_spacing,
  p_vector = c("PCP", "SLR", "RELHUM", "WNDSPD", "TMP_MAX", "TMP_MIN"),
  idw_exponent = 2
)
}
\arguments{
\item{meteo_lst}{A nested list of lists with dataframes.
Nested structure: \code{meteo_lst -> data -> Station ID -> Parameter -> 
  Dataframe (DATE, PARAMETER)},
\code{meteo_lst -> stations -> Dataframe (ID, Name, Elevation, Source, 
  geometry, Long, Lat)}. \cr\cr
meteo_lst can be created using \code{\link{load_template}} function using
'xlsx' template file or it could to be created with \code{\link{load_swat_weather}}
function loading information from SWAT+ model setup weather files.}

\item{catchment_boundary_path}{Character, path to the basin boundary shape file.}

\item{dem_data_path}{Character, path to DEM raster data in the same projection
as the weather station.}

\item{grid_spacing}{Numeric, value for the distance between grid points.
Units of the coordinate system should be used.}

\item{p_vector}{(optional) Character vector representing weather variables to
interpolate. Default is all variables selected
\code{p_vector = c("PCP", "SLR", "RELHUM", "WNDSPD", "TMP_MAX", "TMP_MIN")}.}

\item{idw_exponent}{(optional) Numeric value for the exponent parameter to
be used in interpolation. Default \code{idw_exponent = 2}.}
}
\value{
A nested list of lists with interpolation results.
A nested list of lists with dataframes.
Nested structure: \code{meteo_lst -> data -> Station ID -> Parameter -> 
  Dataframe (DATE, PARAMETER)},
\code{meteo_lst -> stations -> Dataframe (ID, Name, Elevation, Source, 
  geometry, Long, Lat)}.
}
\description{
This function interpolates weather data for a SWAT model and saves results
into nested list format
}
\examples{
\dontrun{
  # Specify paths to weather  data, basing shapeand DEM
  temp_path <- system.file("extdata", "weather_data.xlsx", package = "SWATprepR")
  DEM_path <- system.file("extdata", "GIS/DEM.tif", package = "SWATprepR")
  basin_path <- system.file("extdata", "GIS/basin.shp", package = "SWATprepR")

  # Load weather data template
  met_lst <- load_template(temp_path, 3035)

  # Interpolate and write SWAT model input files
  interpolate(met_lst, basin_path, DEM_path, 2000) 
}
}
\keyword{gap-filling}
