% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{list_to_df}
\alias{list_to_df}
\title{Transforming list to dataframe}
\usage{
list_to_df(lst)
}
\arguments{
\item{lst}{nested list dataframes (for each weather variable)}
}
\value{
dataframe with all collected dataframes in single dataframe.
}
\description{
Transforming list to dataframe
}
\examples{
\dontrun{
temp_path <- system.file("extdata", "weather_data.xlsx", package = "SWATprepR")
met_lst <- load_template(temp_path, 3035)
df <- list_to_df(met_lst$data$ID1)
}
}
\keyword{internal}
