% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading.R
\name{load_calibration_data_template}
\alias{load_calibration_data_template}
\title{Function providing loading of calibration data templates
(data should have been cleaned before).}
\usage{
load_calibration_data_template(template_path, epsg_code = 4326)
}
\arguments{
\item{template_path}{path to *.xlsx file.}

\item{epsg_code}{EPSG code for station coordinates (default 4326 for WGS 84 coordinate system)  system}
}
\value{
list of two dataframes: stations dataframe contains station information,
data contains measurement data
}
\description{
Function providing loading of calibration data templates
(data should have been cleaned before).
}
\examples{
##load_calibration_data_template("templates/calibration_data.xlsx")

}
