% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading.R
\name{load_stations}
\alias{load_stations}
\title{Reading station data from excel templates}
\usage{
load_stations(template_path, epsg_code)
}
\arguments{
\item{template_path}{path to *.xlsx file.}

\item{epsg_code}{EPSG code for station coordinates system.}
}
\value{
sf dataframe with station information.
}
\description{
Reading station data from excel templates
}
\examples{
##read_stations("templates/weather_data.xlsx", 4326)
}
