% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading.R
\name{load_swat_weather2}
\alias{load_swat_weather2}
\title{Loading SWAT+ weather files to R (>20 times faster version)}
\usage{
load_swat_weather2(input_folder)
}
\arguments{
\item{input_folder}{character, path to folder with SWAT+ weather input files (example "my_model").}
}
\value{
nested list of lists with dataframes. Nested structure list -> stations,
list -> Variable -> Dataframe (DATE, VARIABLE).
}
\description{
Loading SWAT+ weather files to R (>20 times faster version)
}
\examples{
\dontrun{
met_lst <- load_swat_weather2("my_folder")
}
}
