% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{min_ks}
\alias{min_ks}
\title{Get min Ks within depth range from SOL_K and SOL_Z}
\usage{
min_ks(df, max_depth)
}
\arguments{
\item{df}{one row dataframe of soil parameters with all SOL_K and SOL_Z columns}

\item{max_depth}{numeric value for depth range from 0.}
}
\value{
numeric min Ks parameter value within depth range
}
\description{
Get min Ks within depth range from SOL_K and SOL_Z
}
\examples{
\dontrun{
df <- data.frame(SOL_K1 = 10, SOL_K2 = 1, SOL_K3 = 2, 
                 SOL_Z1 = 250, SOL_Z2 = 700, SOL_Z3 = 1000)
min_ks(df, 800)
}
}
\keyword{internal}
