% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{my.check.pwd}
\alias{my.check.pwd}
\title{Helper for my.pwd function}
\usage{
my.check.pwd(x)
}
\arguments{
\item{x}{vector of numbers}
}
\value{
logical, if wet day after dry
}
\description{
Helper for my.pwd function
}
\examples{
\dontrun{
my.check.pwd(x)
}
}
\keyword{internal}
