% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{my.check.pww}
\alias{my.check.pww}
\title{Helper for my.pww function}
\usage{
my.check.pww(x)
}
\arguments{
\item{x}{vector of numbers}
}
\value{
logical, if wet day after wet
}
\description{
Helper for my.pww function
}
\examples{
\dontrun{
my.check.pww(x)
}
}
\keyword{internal}
