% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{my.pcpd}
\alias{my.pcpd}
\title{Function to calculate pcp_days parameter}
\usage{
my.pcpd(x, nyears, na.rm = FALSE)
}
\arguments{
\item{x}{vector of numbers}

\item{nyears}{numeric number of years available for PCP variable}

\item{na.rm}{logical, if NA should be cleaned}
}
\value{
numeric result
}
\description{
Function to calculate pcp_days parameter
}
\examples{
\dontrun{
my.pcpd(x)
}
}
\keyword{internal}
