% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{my.pcpmhhr}
\alias{my.pcpmhhr}
\title{Function to calculate hhr parameter}
\usage{
my.pcpmhhr(x, na.rm = FALSE)
}
\arguments{
\item{x}{vector of numbers}

\item{na.rm}{logical, if NA should be cleaned}
}
\value{
numeric result
}
\description{
Function to calculate hhr parameter
}
\examples{
\dontrun{
my.pcpmhhr(x)
}
}
\keyword{internal}
