% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{my.pwd}
\alias{my.pwd}
\title{Function to calculate wet_dry parameter}
\usage{
my.pwd(x, na.rm = FALSE)
}
\arguments{
\item{x}{vector of numbers}

\item{na.rm}{logical, if NA should be cleaned}
}
\value{
numeric result
}
\description{
Function to calculate wet_dry parameter
}
\examples{
\dontrun{
my.pwd(x)
}
}
\keyword{internal}
