% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{my.pww}
\alias{my.pww}
\title{Function to calculate wet_wet parameter}
\usage{
my.pww(x, na.rm = FALSE)
}
\arguments{
\item{x}{vector of numbers}

\item{na.rm}{logical, if NA should be cleaned}
}
\value{
numeric result
}
\description{
Function to calculate wet_wet parameter
}
\examples{
\dontrun{
my.pww(x)
}
}
\keyword{internal}
