% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{my.skew}
\alias{my.skew}
\title{Function to calculate pcp_skew parameter}
\usage{
my.skew(x, na.rm = FALSE)
}
\arguments{
\item{x}{vector of numbers}

\item{na.rm}{logical, if NA should be cleaned}
}
\value{
numeric result
}
\description{
Function to calculate pcp_skew parameter
}
\examples{
\dontrun{
my.skew(x)
}
}
\keyword{internal}
