% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{nyears}
\alias{nyears}
\title{Count number of years data available for variable}
\usage{
nyears(df, col)
}
\arguments{
\item{df}{dataframe with DATE column and at least one variable column.}

\item{col}{variable column name}
}
\value{
numeric number on year for which data is available.
}
\description{
Count number of years data available for variable
}
\examples{
\dontrun{
nyears(df, "PCP")
}
}
\keyword{internal}
