% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_cal_data}
\alias{plot_cal_data}
\title{Plotting calibration data}
\usage{
plot_cal_data(df, stations)
}
\arguments{
\item{df}{dataframe with formatted data (Station, DATE, Variables and Values columns are needed)}

\item{stations}{character vector listing stations, which should be selected for figure.}
}
\value{
plotly object of interactive figure.
}
\description{
Plotting calibration data
}
\examples{
temp_path <- system.file("templates", "calibration_data.xlsx", package = "svatools")
cal_data <- load_template(temp_path)
plot_cal_data(cal_data$data, c("1", "2", "3","10"))
}
