% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_fractions}
\alias{plot_fractions}
\title{Plot regression and fractions for each month between parts of nutrient and total}
\usage{
plot_fractions(df, station, total_var, min_vars)
}
\arguments{
\item{df}{dataframe with formatted data (Station, DATE, Variables and Values columns are needed).}

\item{station}{character vector indicating station/s, which should be selected for figure.}

\item{total_var}{character vector for variable selected to represent total of certain variable.}

\item{min_vars}{character vector for variable/s selected to represent mineral or organic of certain variable.}
}
\value{
list of two ggplot objects: regression - monthly regression plots for total vs some part
(i.e. mineral or organic), fraction - monthly fraction values.
}
\description{
Plot regression and fractions for each month between parts of nutrient and total
}
\examples{
temp_path <- system.file("extdata", "calibration_data.xlsx", package = "SWATprepR")
cal_data <- load_template(temp_path)
plot_fractions(cal_data$data, c("4"), c("NT"), c("N-NO3", "N-NH4", "N-NO2"))
plot_fractions(cal_data$data, c("4"), c("PT"), c("P-PO4"))
}
