% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_many}
\alias{plot_many}
\title{Plotting stations with few data points}
\usage{
plot_many(df, drop_st = c())
}
\arguments{
\item{df}{dataframe with formatted data (Station, DATE, Variables and Values columns are needed)}

\item{drop_st}{option parameter of character vector, which stations should be excluded for figure.
Stations with a lot of data should be plotted with plot_one function.}
}
\value{
plotly object of interactive figure
}
\description{
Plotting stations with few data points
}
\examples{
## plot_many(df)
}
