% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_map}
\alias{plot_map}
\title{Preparing interactive map of monitoring points with data inside}
\usage{
plot_map(df, df_station, rch, shp)
}
\arguments{
\item{df}{dataframe with formatted data (Station, DATE, Variables and Values columns are needed).}

\item{df_station}{dataframe with formatted station data (ID, geometry columns are needed).}

\item{rch}{sf dataframe for reaches (id, geometry columns are needed).}

\item{shp}{sf dataframe for basin (name, geometry columns are needed).}
}
\value{
leaflet object of interactive map with monitoring data opening while pressing on points.
}
\description{
Preparing interactive map of monitoring points with data inside
}
\examples{
temp_path <- system.file("templates", "calibration_data.xlsx", package = "svatools")
cal_data <- load_template(temp_path)
plot_map(cal_data$data, cal_data$stations, rch, shp)
}
