% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_Nmin_NT}
\alias{plot_Nmin_NT}
\title{Plot N min to NT regression in each month and fraction Nmin of NT}
\usage{
plot_Nmin_NT(df, station)
}
\arguments{
\item{df}{dataframe with formatted data (Station, DATE, Variables and Values columns are needed)}

\item{station}{character vector indicating station/s, which should be selected for figure.}
}
\value{
list of two ggplot objects: regression - monthly regression plots for NT vs Nmin,
fraction - monthly fraction values Nmin/NT
}
\description{
Plot N min to NT regression in each month and fraction Nmin of NT
}
\examples{
temp_path <- system.file("templates", "calibration_data.xlsx", package = "svatools")
cal_data <- load_template(temp_path)
plot_Nmin_NT(cal_data$data, station = c("4"))
}
