% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_one}
\alias{plot_one}
\title{Plotting single, data rich station}
\usage{
plot_one(df, station)
}
\arguments{
\item{df}{dataframe with formatted data (Station, DATE, Variables and Values columns are needed)}

\item{station}{character indicating station, which should be selected for figure.}
}
\value{
plotly object of interactive figure.
}
\description{
Plotting single, data rich station
}
\examples{
temp_path <- system.file("templates", "calibration_data.xlsx", package = "svatools")
cal_data <- load_template(temp_path)
plot_one(cal_data$data, station = "4")
}
