% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_Pmin_PT}
\alias{plot_Pmin_PT}
\title{Plot Pmin fraction of PT}
\usage{
plot_Pmin_PT(df, station)
}
\arguments{
\item{df}{dataframe with formatted data (Station, DATE, Variables and Values columns are needed)}

\item{station}{character vector indicating station/s, which should be selected for figure.}
}
\value{
ggplot object for monthly Pmin/PT fraction box plot
}
\description{
Plot Pmin fraction of PT
}
\examples{
temp_path <- system.file("templates", "calibration_data.xlsx", package = "svatools")
cal_data <- load_template(temp_path)
plot_Pmin_PT(cal_data$data, station = unique(station$ID))
}
