% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_ts_fig}
\alias{plot_ts_fig}
\title{Function to prepare static time series figure for single station}
\usage{
plot_ts_fig(station, df)
}
\arguments{
\item{station}{character indicating station, which should be selected for figure.}

\item{df}{dataframe with formatted data (Station, DATE, Variables and Values columns are needed)}
}
\value{
ggplot object of static figure.
}
\description{
Function to prepare static time series figure for single station
}
\examples{
\dontrun{
temp_path <- system.file("extdata", "calibration_data.xlsx", package = "svatools")
cal_data <- load_template(temp_path)
plot_ts_fig("4", cal_data$data)
}
}
\keyword{internal}
