% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_weather}
\alias{plot_weather}
\title{Prepare the plotly figure for weather data}
\usage{
plot_weather(meteo_lst, par, period = "day", fn_summarize = "mean")
}
\arguments{
\item{meteo_lst}{nested list of lists with dataframes.
Nested structure meteo_lst -> data -> Station ID -> Parameter -> Dataframe (DATE, PARAMETER).}

\item{par}{character marking weather variable to extract (i.e. "PCP", "SLR", etc).}

\item{period}{character describing, which time interval to display (default is "day",
other examples are "week", "month", etc).}

\item{fn_summarize}{function to recalculate to time interval (default is "mean", other examples
are "median", "sum", etc).}
}
\value{
plotly figure object with displayed weather data
}
\description{
Prepare the plotly figure for weather data
}
\examples{
temp_path <- system.file("templates", "weather_data.xlsx", package = "svatools")
met_lst <- load_template(temp_path, 4326)
plot_weather(met_lst, "PCP", "month", "sum")
}
