% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_wgn_comparison}
\alias{plot_wgn_comparison}
\title{Plot wgn parameters comparison}
\usage{
plot_wgn_comparison(
  meteo_lst1,
  meteo_lst2,
  station1,
  station2,
  type1 = "set 1",
  type2 = "set 2",
  title = "comparison"
)
}
\arguments{
\item{meteo_lst1}{first nested list of lists with dataframes.
meteo_lst nested list of lists with dataframes.
Nested structure meteo_lst -> data -> Station ID -> Parameter -> Dataframe (DATE, PARAMETER).
Nested meteo_lst -> stations Dataframe (ID, Name, Elevation, Source, geometry, Long, Lat).}

\item{meteo_lst2}{second nested list of lists with dataframes.
meteo_lst nested list of lists with dataframes.
Nested structure meteo_lst -> data -> Station ID -> Parameter -> Dataframe (DATE, PARAMETER).
Nested meteo_lst -> stations Dataframe (ID, Name, Elevation, Source, geometry, Long, Lat).}

\item{station1}{character, id of one station in the first list selected for comparison (example "ID1").}

\item{station2}{character, id of one station in the second list selected for comparison (example "ID1").}

\item{type1}{character, naming of the first dataset (example "measured").}

\item{type2}{character, naming of the second dataset (example "netCDF").}

\item{title}{character, information to be added in figure title}
}
\value{
ggplot figure of bar plots
}
\description{
Plot wgn parameters comparison
}
\examples{
\dontrun{
temp_path <- system.file("extdata", "weather_data.xlsx", package = "svatools")
met_lst <- load_template(temp_path, 4326)
plot_wgn_comparison(met_lst, met_lst, "ID1", "ID2", "Station 1", "Station 2", "some text")
}
}
