% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparing.R
\name{prepare_climate}
\alias{prepare_climate}
\title{Function to prepare or update climate data text input files in SWAT+ model}
\usage{
prepare_climate(meteo_lst, write_path, period_starts = NA, period_ends = NA)
}
\arguments{
\item{meteo_lst}{meteo_lst nested list of lists with dataframes.
Nested structure meteo_lst -> data -> Station ID -> Parameter -> Dataframe (DATE, PARAMETER).
Nested meteo_lst -> stations Dataframe (ID, Name, Elevation, Source, geometry, Long, Lat).}

\item{write_path}{character, path to SWAT+ txtinout folder (example "my_model").}

\item{period_starts}{character, date string (example '1991-01-01'). Optional (\code{default = NA},
stands for all available in data).}

\item{period_ends}{character, date string (example '2020-12-31'). Optional (\code{default = NA},
stands for all available in data).}
}
\value{
Fill or update multiple weather related weather text files.
}
\description{
Function to prepare or update climate data text input files in SWAT+ model
}
\examples{
\dontrun{
prepare_climate(meteo_lst, "output", "1991-01-01", "2020-12-31")
}
}
