% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparing.R
\name{prepare_pt_source}
\alias{prepare_pt_source}
\title{Prepare point source data model text files}
\usage{
prepare_pt_source(pt_lst, project_path, write_path = project_path)
}
\arguments{
\item{pt_lst}{nested list of lists with dataframes.
Nested structure pt_lst -> data -> Dataframe (name,    DATE,    flo, ...)
pt_lst -> st -> Dataframe (name, Lat, Long)}

\item{project_path}{character, path to SWAT+ txtinout folder (example "my_model").}

\item{write_path}{character, path to SWAT+ txtinout folder (example "my_model"). Optional (\code{default = project_path}.}
}
\value{
text files for SWAT+ model in write_path
}
\description{
Prepare point source data model text files
}
\examples{
\dontrun{
temp_path <- system.file("extdata", "pnt_data.xls", package = "SWATprepR")
pnt_data <- load_template(temp_path)
prepare_pt_source(pnt_data, "txtinout")
}
}
