% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparing.R
\name{prepare_wgn}
\alias{prepare_wgn}
\title{Generate Weather Generator (WGN) Data for SWAT+ Model}
\usage{
prepare_wgn(
  meteo_lst,
  TMP_MAX = NULL,
  TMP_MIN = NULL,
  PCP = NULL,
  RELHUM = NULL,
  WNDSPD = NULL,
  MAXHHR = NULL,
  SLR = NULL
)
}
\arguments{
\item{meteo_lst}{Nested list with dataframes.
Nested structure: \code{meteo_lst -> data -> Station ID -> Parameter -> 
  Dataframe (DATE, PARAMETER)},
\code{meteo_lst -> stations -> Dataframe (ID, Name, Elevation, Source, 
  geometry, Long, Lat)}.}

\item{TMP_MAX}{(optional) Dataframe with two columns: DATE : POSIXct, TMP_MAX : num.
This parameter refers to data, which should be used instead if TMP_MAX variable
is missing for a station.  Default \code{TMP_MAX = NULL}, data of the closest
station with data will be used.}

\item{TMP_MIN}{(optional) Dataframe with two columns: DATE : POSIXct, TMP_MIN : num.
This parameter refers to data, which should be used instead if TMP_MIN variable
is missing for a station. Default \code{TMP_MIN = NULL}, indicating that data of the closest
station with data will be used.}

\item{PCP}{(optional) Dataframe with two columns: DATE : POSIXct, PCP : num.
This parameter refers to data, which should be used instead if PCP variable
is missing for a station. Default \code{PCP = NULL}, indicating that data of the closest
station with data will be used.}

\item{RELHUM}{(optional) Dataframe with two columns: DATE : POSIXct, RELHUM : num.
This parameter refers to data, which should be used instead if RELHUM variable
is missing for a station. Default \code{RELHUM = NULL}, indicating that data of the closest
station with data will be used.}

\item{WNDSPD}{(optional) Dataframe with two columns: DATE : POSIXct, WNDSPD : num.
This parameter refers to data, which should be used instead if WNDSPD variable
is missing for a station. Default \code{WNDSPD = NULL}, indicating that data of the closest
station with data will be used.}

\item{MAXHHR}{(optional) Dataframe with two columns: DATE : POSIXct, MAXHHR : num.
This parameter refers to data, which should be used instead if MAXHHR variable
is missing for a station. Default \code{MAXHHR = NULL}, indicating that data of the closest
station with data will be used.}

\item{SLR}{(optional) Dataframe with two columns: DATE : POSIXct, SLR : num.
This parameter refers to data, which should be used instead if SLR variable
is missing for a station. Default \code{SLR = NULL}, indicating that data of the closest
station with data will be used.}
}
\value{
List of two dataframes: wgn_st - WGN station data, wgn_data - WGN data.
}
\description{
This function generates weather generator (WGN) data for a SWAT+ model
based on meteorological data.
}
\examples{
\dontrun{
  # Example usage:
  temp_path <- system.file("extdata", "weather_data.xlsx", package = "SWATprepR")
  met_lst <- load_template(temp_path, 3035)
  # Generate WGN data
  wgn <- prepare_wgn(met_lst,
                     TMP_MAX = met_lst$data$ID10$TMP_MAX,
                     TMP_MIN = met_lst$data$ID10$TMP_MIN,
                     PCP = met_lst$data$ID9$PCP,
                     RELHUM = met_lst$data$ID9$RELHUM,
                     WNDSPD = met_lst$data$ID10$WNDSPD,
                     MAXHHR = met_lst$data$ID11$MAXHHR,
                     SLR = met_lst$data$ID9$SLR)
}
}
\seealso{
\code{\link{load_template}}
}
\keyword{parameters}
