% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rch}
\alias{rch}
\title{River data}
\format{
A sf data.frame with 2 variables:
\describe{
\item{id}{id of each reach.}
\item{geometry}{Geometry of objects.}
}
}
\usage{
rch
}
\description{
Example GIS river dataset for Polish case study
}
\keyword{datasets}
