% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading.R
\name{read_tbl}
\alias{read_tbl}
\title{Loading SWAT+ text files into dataframe}
\usage{
read_tbl(tbl_name, proj_path, row_data_start, row_col_names)
}
\arguments{
\item{tbl_name}{character, name of the file to be read eaxmple ('rout_unit.con').}

\item{proj_path}{character, path to SWAT+ txtinout folder (example "my_model").}

\item{row_data_start}{numeric, row number from which data are being written.}

\item{row_col_names}{numeric, row nu,mber in which column names are being written.}
}
\value{
dataframe with information from text file
}
\description{
Loading SWAT+ text files into dataframe
}
\examples{
\dontrun{
df <- read_tbl('rout_unit.con', 'model_folder', 3, 2) 
}
}
