% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shp}
\alias{shp}
\title{Basin boundary data}
\format{
A data.frame with 2 variables:
\describe{
\item{name}{Name of the catchment.}
\item{geometry}{Geometry of objects.}
}
}
\usage{
shp
}
\description{
Example basin GIS dataset for Polish case study
}
\keyword{datasets}
