% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading.R
\name{sol_to_usersoil}
\alias{sol_to_usersoil}
\title{Convert SWAT+ soils.sol data to usersoil format}
\usage{
sol_to_usersoil(project_path)
}
\arguments{
\item{project_path}{Character, path to the SWAT+ project folder (example "my_model").}
}
\value{
A data frame representing user soil data.
}
\description{
This function reads the soils.sol file from a SWAT+ project folder and converts
the data into usersoil format to examine/update with other functions.
}
\examples{
\dontrun{
usersoil <- sol_to_usersoil("my_model")
}
}
\seealso{
\code{\link{get_usersoil_table}}, \code{\link{usersoil_to_sol}}
}
\keyword{loading}
