% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{sp.na.omit}
\alias{sp.na.omit}
\title{Function to remove NAs in sp dataframe}
\usage{
sp.na.omit(x, margin = 1)
}
\arguments{
\item{x}{sp dataframe}

\item{margin}{numeric. Optional, default 1 for remove rows, 2 for removing columns.}
}
\value{
sp dataframe without NAs in data
}
\description{
Function to remove NAs in sp dataframe
}
\examples{
\dontrun{
sp_df <- sp.na.omit(sp_df)
}
}
\keyword{internal}
