% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{transform_to_list}
\alias{transform_to_list}
\title{Transforming list of sp dataframes to list of list}
\usage{
transform_to_list(list_sp, start_date, end_date)
}
\arguments{
\item{list_sp}{a list of SpatialPointsDataFrames (for each weather variable)
prepared with  \code{\link{interpolate}} function.}

\item{start_date}{time series starting date string. Example "1998-01-01".}

\item{end_date}{time series ending date string. Example "2021-12-31".}
}
\value{
meteo_lst nested list of lists with dataframes.
Nested structure meteo_lst -> data -> Station ID -> Parameter -> Dataframe (DATE, PARAMETER).
Nested meteo_lst -> stations Dataframe (ID, Name, Elevation, Source, geometry, Long, Lat).
}
\description{
Transforming list of sp dataframes to list of list
}
\examples{
\dontrun{
###First step to get interpolation results
temp_path <- system.file("extdata", "weather_data.xlsx", package = "SWATprepR")
DEM_path <- system.file("extdata", "GIS/DEM.tif", package = "SWATprepR")
basin_path <- system.file("extdata", "GIS/basin.shp", package = "SWATprepR")
met_lst <- load_template(temp_path, 3035)
result <- interpolate(met_lst, "./output/",  basin_path, DEM_path, 2000) 

###Second step converting interpolation results to list of list format.
start_date <- SWATprepR:::get_dates(met_lst)$min_date
end_date <- SWATprepR:::get_dates(met_lst)$max_date
int_met_lst <- transform_to_list(result, start_date, end_date)
}
}
\keyword{internal}
