% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{update_wst_id}
\alias{update_wst_id}
\title{Updating sqlite database table with wst_id information}
\usage{
update_wst_id(tname, db_path, wst_cli)
}
\arguments{
\item{tname}{character of table name (example "hru_con")}

\item{db_path}{character to sqlite database (example "./output/project.sqlite")}

\item{wst_cli}{data.frame weather_sta_cli table}
}
\value{
updates sql database with table amended with wst_id (id of closest weather station)
}
\description{
Updating sqlite database table with wst_id information
}
\examples{
\dontrun{
update_wst_id("hru_con", "./output/project.sqlite", weather_sta_cli)
}
}
\keyword{internal}
