% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{update_wst_txt}
\alias{update_wst_txt}
\title{Updating wst_id information in model text files}
\usage{
update_wst_txt(fname, write_path, wst_sf, spacing, folder_to_save = "temp")
}
\arguments{
\item{fname}{character, model file name (example 'aquifer.con').}

\item{write_path}{character, path to SWAT+ txtinout folder (example "my_model").}

\item{wst_sf}{sf dataframe with weather station information. Should contain "name" and "geometry" columns.}

\item{spacing}{character, should contain information about spacing between columns in the file.
Example c('\%8s', '\%-12s', rep('\%12s', 5), '\%8s', '\%16s', rep('\%8s', 4)).}

\item{folder_to_save}{character, folder to save function results. Optional, \code{default = "temp"}.}
}
\value{
Updated text file, which would have updated station names
(with nearest stations to features).
}
\description{
Updating wst_id information in model text files
}
\examples{
\dontrun{
spacing <- c('\%8s', '\%-12s', rep('\%12s', 5), '\%8s', '\%16s', 
rep('\%8s', 4), '\%12s', '\%8s', rep('\%12s', 2))
update_wst_txt("reservoir.con", "model_folder", wst_sf, spacing)
}
}
