% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{wbalance_table}
\alias{wbalance_table}
\title{Prepare water balance table}
\usage{
wbalance_table(project_path)
}
\arguments{
\item{project_path}{The path to the directory containing SWAT+ output files.}
}
\value{
A data frame containing water balance parameters
}
\description{
This function reads data from 'basin_wb_aa.txt' and 'basin_aqu_aa.txt'
SWAT+ model output files and prepares various water balance parameters table.
}
\examples{
\dontrun{
df <- wbalance_table("path/to/project")
write.csv(df, "water_balance.csv", row.names = FALSE, quote = FALSE)
}
}
