% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writing.R
\name{write_atmo_cli}
\alias{write_atmo_cli}
\title{Write 'atmo.cli' file}
\usage{
write_atmo_cli(df, write_path, t_ext = "year")
}
\arguments{
\item{df}{dafaframe with "DATE", "NH4_RF", "NO3_RF" , "NH4_DRY"  and "NO3_DRY" columns obtained from \code{\link{get_atmo_dep}}) function.}

\item{write_path}{path to folder where results should be written.}

\item{t_ext}{string, how to prepare file: 'year' for yearly averages, 'month' - monthly averages
and 'annual' for average of all period. Optional (default - "year"). If df is with monthly data, only 'month' should be used.}
}
\value{
'atmo.cli' file with one station for catchment.
}
\description{
Write 'atmo.cli' file
}
\examples{
\dontrun{
basin_path <- system.file("extdata", "GIS/basin.shp", package = "svatools")
df <- get_atmo_dep(basin_path)
write_atmo_cli(df, "./output/")
}
}
