% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writing.R
\name{write_input_files_tmp}
\alias{write_input_files_tmp}
\title{Writing input files for TMP data}
\usage{
write_input_files_tmp(write_path, sp_df_mx, sp_df_mn, meteo_lst)
}
\arguments{
\item{write_path}{path to folder where results should be written.}

\item{sp_df_mx}{SpatialPointsDataFrame with resulting interpolated data for TMP_MAX variable.}

\item{sp_df_mn}{SpatialPointsDataFrame with resulting interpolated data for TMP_MIN variable.}

\item{meteo_lst}{nested list of lists with dataframes.}
}
\value{
temperature weather data text files for virtual stations (created during interpolation)
in format usable by the SWAT model.
}
\description{
Writing input files for TMP data
}
\examples{
\dontrun{
write_input_files_tmp("./output/", sp_df_mx, sp_df_mn)
}
}
