% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writing.R
\name{write_ref_file}
\alias{write_ref_file}
\title{Writing reference file for weather data}
\usage{
write_ref_file(write_path, sp_df, par)
}
\arguments{
\item{write_path}{path to folder where results should be written.}

\item{sp_df}{path to folder where results should be written.}

\item{par}{weather variable (i.e. "PCP", "SLR", etc).}
}
\value{
weather virtual stations (created during interpolation) reference file
in format usable by the SWAT model.
}
\description{
Writing reference file for weather data
}
\examples{
\dontrun{
write_ref_file("./output/", sp_df, "PCP")
}
}
