% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writing.R
\name{write_weather_template}
\alias{write_weather_template}
\title{Write weather meteo_lst into template .xlsx file}
\usage{
write_weather_template(
  meteo_lst,
  write_path = "",
  f_name = "weather_data.xlsx"
)
}
\arguments{
\item{meteo_lst}{nested list of lists with dataframes.
Nested structure meteo_lst -> data -> Station ID -> Parameter -> Dataframe (DATE, PARAMETER).}

\item{write_path}{(optional) character path to folder where results should be written (default "").}

\item{f_name}{(optional) character name of the file (default 'weather_data.xlsx').}
}
\value{
.xlsx file written in template format, which could de read by \code{\link{load_template}} function.
}
\description{
Write weather meteo_lst into template .xlsx file
}
\examples{
\dontrun{
write_weather_template(met_lst)
}
}
