% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writing.R
\name{write_wgnmaker_files}
\alias{write_wgnmaker_files}
\title{Preparing input files for WGNmaker.}
\usage{
write_wgnmaker_files(write_path, meteo_lst, wgn_stations_lst, station_name)
}
\arguments{
\item{write_path}{path to folder where results should be written.}

\item{meteo_lst}{nested list of lists with dataframes.
Nested structure meteo_lst -> data -> Station ID -> Parameter -> Dataframe (DATE, PARAMETER).}

\item{wgn_stations_lst}{List of selected stations for each parameter to be used in data preparation.
Example list(PCP = "ID12",  SLR = "ID11", RELHUM = "ID12", TMP_MAX = "ID12",
TMP_MIN = "ID12", WNDSPD = "ID12", MAXHHR = "ID12")}

\item{station_name}{string with name of station.}
}
\value{
Excel files needed for WGNmaker
}
\description{
WGNmaker excel macro could be downloaded from following link: https://swat.tamu.edu/media/41583/wgen-excel.zip
}
\examples{
\dontrun{
write_wgnmaker_files("./wgn/", meteo_lst, wgn_stations_lst, "Station1")
}
}
