% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparing.R
\name{add_atmo_dep}
\alias{add_atmo_dep}
\title{Update SWAT+ project with atmospheric deposition data}
\usage{
add_atmo_dep(df, write_path, t_ext = "year")
}
\arguments{
\item{df}{A data frame containing columns "DATE," "NH4_RF," "NO3_RF,"
"NH4_DRY," and "NO3_DRY" obtained from the \code{\link{get_atmo_dep}}
function.}

\item{write_path}{Character, path to the SWAT+ txtinout folder (example "my_model").}

\item{t_ext}{(optional) A string indicating the type of time aggregation:
'year' for yearly averages, 'month' for monthly averages, and 'annual' for
the average of the entire period. Default default \code{t_ext = "year"}.}
}
\value{
Writes data to 'atmodep.cli' file and can update the SWAT+ SQLite
database with codes and connections to atmospheric deposition data.
}
\description{
This function updates a SWAT+ model text files and writes the \code{atmodep.cli}
file for a given dataframe containing atmospheric deposition data.
}
\examples{
\dontrun{
  basin_path <- system.file("extdata", "GIS/basin.shp", package = "SWATprepR")
  df <- get_atmo_dep(basin_path)
  project_path <- "./output/test"
  add_atmo_dep(df, project_path)
}
}
\seealso{
\code{\link{get_atmo_dep}}
}
\author{
Svajunas Plunge (svajunas.plunge@gmail.com)

Moritz Shore (moritzshore@gmail.com)
}
\keyword{writing}
