% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparing.R
\name{add_atmo_dep}
\alias{add_atmo_dep}
\title{Update SQLite database with atmospheric deposition data}
\usage{
add_atmo_dep(df, db_path, t_ext = "year")
}
\arguments{
\item{df}{A data frame containing columns "DATE," "NH4_RF," "NO3_RF,"
"NH4_DRY," and "NO3_DRY" obtained from the \code{\link{get_atmo_dep}} function.}

\item{db_path}{A character string representing the path to the SQLite
database (e.g., "./output/project.sqlite").}

\item{t_ext}{(optional) A string indicating the type of time aggregation: 'year' for
yearly averages, 'month' for monthly averages, and 'annual' for the average
of the entire period. Default default \code{t_ext = "year"}.}
}
\value{
Writes data to 'atmodep.cli' file and updates the SQLite database
with codes and connections to atmospheric deposition data.
}
\description{
This function updates an SQLite database with atmospheric deposition data.
}
\examples{
\dontrun{
  basin_path <- system.file("extdata", "GIS/basin.shp", package = "SWATprepR")
  df <- get_atmo_dep(basin_path)
  db_path <- "./output/test/project.sqlite"
  add_atmo_dep(df, db_path)
}
}
\seealso{
\code{\link{get_atmo_dep}}
}
\keyword{writing}
