% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparing.R
\name{add_atmo_dep}
\alias{add_atmo_dep}
\title{Update sqlite database with atmospheric deposition data}
\usage{
add_atmo_dep(db_path, t_ext = "year")
}
\arguments{
\item{db_path}{character to sqlite database (example "./output/project.sqlite")}

\item{t_ext}{string, 'year' for yearly averages, 'month' - monthly averages
and 'annual' for average of all period. Optional (default - "year").}
}
\value{
updated sqlite database with codes and connection to atmospheric deposition  data in 'atmo.cli' file
}
\description{
Update sqlite database with atmospheric deposition data
}
\examples{
\dontrun{
db_path <- "./output/test/project.sqlite"
add_atmo_dep(db_path)
}
}
