% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{clean_outliers}
\alias{clean_outliers}
\title{Function to clean outliers outside by some standard deviations}
\usage{
clean_outliers(df, times_sd = 3)
}
\arguments{
\item{df}{dataframe with columns c("Station", "DATE", "Variables", "Values", "Source").}

\item{times_sd}{numeric value representing multiplication factor for standard deviation.
Values outside mean - sd X times_sd, mean + sd X times_sd are identified as outliers.
Optional with default value 3.}
}
\value{
list of two dataframes. \emph{newdf} dataframe contains dataframe cleaned from outliers.
\emph{dropped}  dataframe contains data, which was removed from \emph{newdf} dataframe.
}
\description{
Function to clean outliers outside by some standard deviations
}
\examples{
temp_path <- system.file("extdata", "calibration_data.xlsx", package = "SWATprepR")
cal_data <- load_template(temp_path)
lst <- clean_outliers(cal_data$data)
##Looking at data to be removed
print(head(lst$dropped))
##Updating data
cal_data$data <- lst$newdf
}
