% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{clean_wq}
\alias{clean_wq}
\title{Clean water quality data from most typical issues}
\usage{
clean_wq(df, zero_to_min = 1)
}
\arguments{
\item{df}{dataframe with water quality data with  with columns c("Station", "DATE", "Variables", "Values", "Source").}

\item{zero_to_min}{numeric coefficient to zeros by min variable value X zero_to_min. Optional, default 1.}
}
\value{
cleaned dataframe
}
\description{
Clean water quality data from most typical issues
}
\examples{
temp_path <- system.file("extdata", "calibration_data.xlsx", package = "svatools")
cal_data <- load_template(temp_path)
cal_data$data <- clean_wq(cal_data$data)
}
