% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helping.R
\name{df_t}
\alias{df_t}
\title{Transforming sp dataframe to dataframe}
\usage{
df_t(sp_df)
}
\arguments{
\item{sp_df}{sp dataframe}
}
\value{
Dataframe prepared for being written out.
}
\description{
Transforming sp dataframe to dataframe
}
\examples{
##df_t(sp_df)
}
\keyword{internal}
