% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparing.R
\name{extract_rotation}
\alias{extract_rotation}
\title{Extract rotation information from raster file}
\usage{
extract_rotation(df, start_year, tif_name, r_path, lookup, lu_constant = c())
}
\arguments{
\item{df}{An sf data.frame with land use. Columns "id" and "type" should be
present.}

\item{start_year}{A numeric value representing the year from which data begins.}

\item{tif_name}{A string for the name of the .tif raster file.}

\item{r_path}{A string for the path to the .tif file.}

\item{lookup}{A dataframe with a "lc1" column for numeric codes and a "type"
column for text.}

\item{lu_constant}{(optional) A vector of strings with land uses to be kept constant in
land use (e.g., water, urban areas). Default \code{lu_constant = c()}.}
}
\value{
An sf data.frame with land use amended with crop rotation information.
}
\description{
This function extracts crop rotation information from a raster file and
amends the land use data accordingly.
}
\examples{
\dontrun{
  library(sf)
  
  # Loading land use/crop layer
  lu_path <- system.file("extdata", "GIS/lu_layer.shp", package = "SWATprepR")
  lu <- st_read(lu_path, quiet = TRUE) \%>\% mutate(id = row_number())
  
  # Preparing lookup table
  lookup <- data.frame(lc1 = seq(1:length(unique(c(lu$type)))), 
  type = unique(c(lu$type)))
  lu_constant <- c("fesc", "orch", "frst", "frse", "frsd", "urld", "urhd", 
  "wetl", "past", "watr", "agrl")

  # Extracting rotation information from raster
  # Raster information should have been prepared with remote sensing 
  lu_rot <- extract_rotation(lu, 2015, "cropmaps.tif", "./output/", lookup, 
  lu_constant)
}
}
\references{
Mészáros, J., & Szabó, B. (2022). Script to derive and apply crop
classification based on Sentinel 1 satellite radar images in Google Earth
Engine platform. \url{https://doi.org/10.5281/zenodo.6700122}
}
\keyword{remote-sensing}
