% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparing.R
\name{fill_with_closest}
\alias{fill_with_closest}
\title{Filling missing variables from the closest stations, which has data.}
\usage{
fill_with_closest(
  meteo_lst,
  par_fill = c("TMP_MAX", "TMP_MIN", "PCP", "RELHUM", "WNDSPD", "SLR")
)
}
\arguments{
\item{meteo_lst}{meteo_lst nested list of lists with dataframes.
Nested structure meteo_lst -> data -> Station ID -> Parameter -> Dataframe (DATE, PARAMETER).
Nested meteo_lst -> stations Dataframe (ID, Name, Elevation, Source, geometry, Long, Lat).}

\item{par_fill}{vector of variables to be filled. Optional (default is c("TMP_MAX", "TMP_MIN","PCP", "RELHUM", "WNDSPD", "SLR")).}
}
\value{
list of dataframe with filled data. Returned list is for meteo_lst$data.
}
\description{
Filling missing variables from the closest stations, which has data.
}
\examples{
\dontrun{
temp_path <- system.file("extdata", "weather_data.xlsx", package = "SWATprepR")
met_lst <- load_template(temp_path, 3035)
met_lst$data <- fill_with_closest(met_lst, c("TMP_MAX", "TMP_MIN"))
}
}
