% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparing.R
\name{get_atmo_dep}
\alias{get_atmo_dep}
\title{Extract EMEP atmospheric deposition data for a catchment}
\usage{
get_atmo_dep(
  catchment_boundary_path,
  t_ext = "year",
  start_year = 1990,
  end_year = 2020
)
}
\arguments{
\item{catchment_boundary_path}{path to basin boundary shape file.}

\item{t_ext}{string, which EMEP data to access 'year' for yearly averages, 'month' - monthly averages. Optional (default - 'year').}

\item{start_year}{integer year to start data extraction. Optional (default - 1990).}

\item{end_year}{integer year to end data extraction. Optional (default - 2020).}
}
\value{
dafaframe with "DATE", "NH4_RF", "NO3_RF" , "NH4_DRY"  and "NO3_DRY" columns. Values in SWAT+ units.
"NH4_RF" - ammonia in rainfall (mg/l), "NO3_RF" - nitrate in rainfall (mg/l),
NH4_RF - ammonia deposition (kg/ha/yr), "NO3_DRY" - nitrate dry deposition (kg/ha/yr)
}
\description{
Extract EMEP atmospheric deposition data for a catchment
}
\examples{
\dontrun{
basin_path <- system.file("extdata", "GIS/basin.shp", package = "svatools")
df <- get_atmo_dep(basin_path)
##Plot results
ggplot(pivot_longer(df, !DATE, names_to = "par", values_to = "values"), aes(x = DATE, y = values))+
geom_line()+
facet_wrap(~par, scales = "free_y")+
theme_bw()
}
}
