% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading.R
\name{get_atmo_dep}
\alias{get_atmo_dep}
\title{Extract EMEP atmospheric deposition data for a catchment}
\usage{
get_atmo_dep(
  catchment_boundary_path,
  t_ext = "year",
  start_year = 1990,
  end_year = 2022
)
}
\arguments{
\item{catchment_boundary_path}{Path to the basin boundary shape file.}

\item{t_ext}{(optional) String, specifying the EMEP data to access.
'year' for yearly averages, 'month' for monthly averages. Default
\code{t_ext = 'year'}.}

\item{start_year}{(optional) Integer representing the starting year for data
extraction. Default \code{start_year = 1990}.}

\item{end_year}{Integer representing the ending year for data extraction.
Default \code{end_year = 2022}.}
}
\value{
A dataframe with columns "DATE", "NH4_RF", "NO3_RF", "NH4_DRY", and
"NO3_DRY".
Values are in SWAT+ units.
"NH4_RF" - ammonia nitrogen in rainfall (mg/l),
"NO3_RF" - nitrate nitrogen in rainfall (mg/l),
NH4_DRY - ammonia nitrogen dry deposition (kg/ha/yr),
"NO3_DRY" - nitrate nitrogen dry deposition (kg/ha/yr).
}
\description{
This function extracts EMEP atmospheric deposition data averaged for
a specified catchment.
}
\examples{
\dontrun{
  # Specify the path to the basin boundary shape file
  catchment_boundary_path <- system.file("extdata", "GIS/basin.shp", package = "SWATprepR")
  
  # Get atmospheric deposition data for the catchment
  df <- get_atmo_dep(catchment_boundary_path)
  
  # Plot results
  ggplot(pivot_longer(df, !DATE, names_to = "par", values_to = "values"), 
  aes(x = DATE, y = values))+
    geom_line()+
    facet_wrap(~par, scales = "free_y")+
    theme_bw()
}
}
\references{
See the EMEP website for more information: \url{https://www.emep.int/mscw/mscw_moddata.html}
}
\seealso{
Please read about SWAT+ atmospheric input data on \url{https://swatplus.gitbook.io/io-docs/introduction/climate/atmo.cli}.
}
\keyword{loading}
