% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helping.R
\name{get_grid}
\alias{get_grid}
\title{Prepare grid for interpolation}
\usage{
get_grid(sp_df, grid_spacing)
}
\arguments{
\item{sp_df}{SpatialPointsDataFrame with columns as days and rows as stations. Lat and Long columns should be included.}

\item{grid_spacing}{Grid spacing. Depending on coordinate system could be in meters, or degrees.}
}
\value{
Grid needed for the interpolation.
}
\description{
Prepare grid for interpolation
}
\examples{
##get_grid(sp_sf, 2000)
}
