% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparing.R
\name{get_interpolated_data}
\alias{get_interpolated_data}
\title{Main interpolation function}
\usage{
get_interpolated_data(
  meteo_lst,
  par,
  catchment_boundary_path,
  dem_data_path,
  grid_spacing,
  idw_exponent = 2
)
}
\arguments{
\item{meteo_lst}{nested list of lists with dataframes.
Nested structure meteo_lst -> data -> Station ID -> Parameter -> Dataframe (DATE, PARAMETER).}

\item{par}{character representing weather variable to extract (i.e. "PCP", "SLR", etc).}

\item{catchment_boundary_path}{path to basin boundary shape file.}

\item{dem_data_path}{path to DEM raster data in same projection as weather station.}

\item{grid_spacing}{numeric value for distance in grid. Units of coordinate system should be used.}

\item{idw_exponent}{numeric value for exponent parameter to be used in interpolation.
(optional, default value is 2).}
}
\value{
SpatialPointsDataFrame with interpolated data.
}
\description{
Main interpolation function
}
\examples{
\dontrun{
temp_path <- system.file("extdata", "weather_data.xlsx", package = "svatools")
DEM_path <- system.file("extdata", "GIS/DEM.tif", package = "svatools")
basin_path <- system.file("extdata", "GIS/basin.shp", package = "svatools")
met_lst <- load_template(temp_path, 3035)
get_interpolated_data(met_lst, "PCP", basin_path, DEM_path, 2000, 2)
}
}
