% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparing.R
\name{get_interpolated_data}
\alias{get_interpolated_data}
\title{Main interpolation function}
\usage{
get_interpolated_data(meteo_lst, grd, par, shp, dem_data_path, idw_exponent)
}
\arguments{
\item{meteo_lst}{nested list of lists with dataframes.
Nested structure meteo_lst -> data -> Station ID -> Parameter -> Dataframe (DATE, PARAMETER).}

\item{grd}{sp SpatialGrid grid for the interpolation.}

\item{par}{character representing weather variable to extract (i.e. "PCP", "SLR", etc).}

\item{shp}{sf dataframe for defining basin boundary shape.}

\item{dem_data_path}{path to DEM raster data in same projection as weather station.}

\item{idw_exponent}{numeric value for exponent parameter to be used in interpolation.
(optional, default value is 2).}
}
\value{
SpatialPointsDataFrame with interpolated data.
}
\description{
Main interpolation function
}
\examples{
\dontrun{
temp_path <- system.file("extdata", "weather_data.xlsx", package = "SWATprepR")
DEM_path <- system.file("extdata", "GIS/DEM.tif", package = "SWATprepR")
basin_path <- system.file("extdata", "GIS/basin.shp", package = "SWATprepR")
met_lst <- load_template(temp_path, 3035)
get_interpolated_data(met_lst, grd, "PCP", read_sf(basin_path), DEM_path, 2)
}
}
\keyword{internal}
