% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparing.R
\name{get_soil_parameters}
\alias{get_soil_parameters}
\title{Function to fill soil parameters}
\usage{
get_soil_parameters(template_path)
}
\arguments{
\item{template_path}{path to *.xlsx file with at least SNAM    NLAYERS    columns and SOL_Z    CLAY
SILT    SAND    OC columns for each available soil layer.}
}
\value{
dataframe with fully formatted and filled table of soil parameters for SWAT model.
}
\description{
Function to fill soil parameters
}
\examples{
\dontrun{
temp_path <- system.file("templates", "soil_parameters.xlsx", package = "svatools")
library(euptf2)
soil <- get_soil_parameters(temp_path)
str(soil)
}
}
