% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparing.R
\name{get_soil_parameters}
\alias{get_soil_parameters}
\title{Function to fill soil parameters}
\usage{
get_soil_parameters(soilp)
}
\arguments{
\item{soilp}{dataframe with SNAM,    NLAYERS    columns and SOL_Z, CLAY, SILT, SAND,    OC columns for each available soil layer.}
}
\value{
dataframe with fully formatted and filled table of soil parameters for SWAT model.
}
\description{
Function to fill soil parameters
}
\examples{
\dontrun{
soil <- get_soil_parameters(temp_path)
}
}
\keyword{internal}
