% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparing.R
\name{get_usersoil_table}
\alias{get_usersoil_table}
\title{Prepare user soil table for SWAT model}
\usage{
get_usersoil_table(csv_path, hsg = FALSE, keep_values = FALSE, nb_lyr = NA)
}
\arguments{
\item{csv_path}{Character path to the CSV file (e.g., "usersoil_lrew.csv").
The file should be comma-separated with minimum columns of:
\itemize{
\item SNAM - name of soil type,
\item NLAYERS - number of soil layers in the soil, and for each soil layer i (minimum 1).
\item SOL_Zi - depth from soil surface to bottom of layer in mm;
\item CLAYi - clay content (particles <0.002 mm) in \% soil weight;
\item SILTi - silt content (particles between 0.002 and 0.05 mm) in \% soil weight;
\item SANDi - sand content (particles between 0.05 and 2 mm) in \% soil weight;
\item SOL_CBNi - organic carbon content (\% soil weight) for each available soil layer.
}}

\item{hsg}{(optional) Logical, TRUE - prepare soil hydrological groups,
FALSE - no soil hydrological group preparation will be done.
Default \code{ hsg = FALSE}. If \code{hsg = TRUE}, three additional columns
should be in an input table: \cr
\itemize{
\item 'Impervious' - depth to impervious layer (allowed values are "<50cm",
"50-100cm", ">100cm");
\item 'Depth' - depth to water table  (allowed values are "<60cm", "60-100cm",
">100cm");
\item 'Drained' - whether soil is drained (allowed values are "Y" for drained
areas, "N" for areas without working tile drains). \cr
More information can be found in the SWAT+ modeling protocol
\href{https://doi.org/10.5281/zenodo.7463395}{Table 3.3}.
}}

\item{keep_values}{(optional) Logical or character vector, TRUE - keep old values (new
values only will be left where 0 or NA values are present in an input table),
FALSE - keep only new values. A character vector can also be used to keep only
specified columns. For instance, c("HYDGRP", "ROCK1") would keep values of
soil hydro groups and rock data for the first layer, while c("HYDGRP", "ROCK")
would keep values in all rock data columns. Default \code{keep_values = FALSE}.}

\item{nb_lyr}{(optional) Integer, the number of layers resulting user soil
data should contain. Default \code{nb_lyr = NA}, which stands for the
same number as in the input data.}
}
\value{
A dataframe with a fully formatted and filled table of soil parameters
for the SWAT model.
}
\description{
This function prepares a user soil table for the SWAT model based on the
provided CSV file.
}
\examples{
\dontrun{
  usersoils <- get_usersoil_table("table.csv")
  write.csv(usersoils, "usersoils.csv", row.names=FALSE, quote=FALSE)
}
}
\references{
SWAT+ Modeling Protocol: Soil Physical Data Chapter

This function utilizes the PTF functions and methods described in pages 82-91.
For detailed information, refer to: \url{https://doi.org/10.5281/zenodo.7463395}
}
\seealso{
This function requires the euptf2 package.
Please read more how to install, use and what is in it on \url{https://github.com/tkdweber/euptf2}.
}
\keyword{parameters}
