% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading.R
\name{load_template}
\alias{load_template}
\title{Function providing loading for data templates (with single and multiple data sheets)
(data should have been cleaned before).}
\usage{
load_template(template_path, epsg_code = 4326)
}
\arguments{
\item{template_path}{path to *.xlsx file.}

\item{epsg_code}{EPSG code for station coordinates (default 4326 for WGS 84 coordinate system)  system}
}
\value{
list of two: stations dataframe, which contains station information,
second member of list contains contains measurement data in dataframe (if data is in one sheet)
or nested list of dataframes with each parameter in separate dataframe.
}
\description{
Function providing loading for data templates (with single and multiple data sheets)
(data should have been cleaned before).
}
\examples{
##load__template("templates/calibration_data.xlsx")
}
