% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_cal_data}
\alias{plot_cal_data}
\title{Plot Calibration Data}
\usage{
plot_cal_data(df, stations = NULL, variables = NULL)
}
\arguments{
\item{df}{Dataframe with formatted data (requires "Station", "DATE", "
Variables", and "Values" columns). Data can be loaded with
\code{\link{load_template}} function using 'xlsx' template file.}

\item{stations}{Character vector listing stations to be selected for the figure.}

\item{variables}{(Optional) Character vector specifying which parameters
should be included in the figure. Default \code{variables = NULL}, all available
variables will be plotted.}
}
\value{
Plotly object of an interactive figure.
}
\description{
This function generates an interactive plot of calibration data based on the
provided data.
}
\examples{
\dontrun{
# Example using calibration data
temp_path <- system.file("extdata", "calibration_data.xlsx", package = "SWATprepR")
cal_data <- load_template(temp_path)
plot_cal_data(cal_data$data, stations = c("1", "2", "3", "10"), variables = c("PT", "NT"))
}
}
\keyword{plotting}
