% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_cal_data}
\alias{plot_cal_data}
\title{Plotting calibration data}
\usage{
plot_cal_data(df, stations, variables = levels(as.factor(df$Variables)))
}
\arguments{
\item{df}{dataframe with formatted data (Station, DATE, Variables and Values columns are needed)}

\item{stations}{character vector listing stations, which should be selected for figure.}

\item{variables}{optional parameter of character vector, which parameters should be in figure.}
}
\value{
plotly object of interactive figure.
}
\description{
Plotting calibration data
}
\examples{
temp_path <- system.file("extdata", "calibration_data.xlsx", package = "svatools")
cal_data <- load_template(temp_path)
plot_cal_data(cal_data$data, stations = c("1", "2", "3","10"), variables =
c("PT", "Q"))
}
