% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_fractions}
\alias{plot_fractions}
\title{Plot Regression and Fractions for Each Month}
\usage{
plot_fractions(df, station, total_var, min_vars)
}
\arguments{
\item{df}{Dataframe with formatted data (requires "Station", "DATE",
"Variables", and "Values" columns).}

\item{station}{Character vector indicating the station/s to be selected for
the figure.}

\item{total_var}{Character vector for the variable selected to represent the
total of a certain nutrient.}

\item{min_vars}{Character vector for the variable/s selected to represent the
mineral or organic fraction of a certain nutrient.}
}
\value{
A list of two ggplot objects: "regression" for monthly regression
plots and "fraction" for monthly fraction values.
}
\description{
This function generates two sets of plots for monthly regression and fractions
between nutrient parts and the total.
}
\examples{
# Example using calibration data
temp_path <- system.file("extdata", "calibration_data.xlsx", package = "SWATprepR")
cal_data <- load_template(temp_path)
plot_fractions(cal_data$data, c("4"), c("NT"), c("N-NO3", "N-NH4", "N-NO2"))
plot_fractions(cal_data$data, c("4"), c("PT"), c("P-PO4"))
}
\keyword{plotting}
