% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_many}
\alias{plot_many}
\title{Plotting stations with few data points}
\usage{
plot_many(df, drop_st = c())
}
\arguments{
\item{df}{dataframe with formatted data (Station, DATE, Variables and Values columns are needed)}

\item{drop_st}{optional parameter of character vector, which stations should be excluded for figure.
Stations with a lot of data should be plotted with plot_one function.}
}
\value{
plotly object of interactive figure
}
\description{
Plotting stations with few data points
}
\examples{
temp_path <- system.file("templates", "calibration_data.xlsx", package = "svatools")
cal_data <- load_template(temp_path)
plot_many(cal_data$data, drop_st = c("4"))
}
