% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_map}
\alias{plot_map}
\title{Prepare Interactive Map of Monitoring Points}
\usage{
plot_map(df, df_station, rch, shp)
}
\arguments{
\item{df}{Dataframe with formatted data (requires "Station", "DATE",
"Variables", and "Values" columns). Data can be loaded with
\code{\link{load_template}} function using 'xlsx' template file.}

\item{df_station}{Dataframe with formatted station data (requires "ID" and
"geometry" columns).}

\item{rch}{sf dataframe for reaches (requires "id" and "geometry" columns).}

\item{shp}{sf dataframe for basin (requires "name" and "geometry" columns).}
}
\value{
Leaflet object of an interactive map with monitoring data displayed
when points are clicked.
}
\description{
This function generates an interactive map with monitoring points,
allowing users to click on points to view data.
}
\examples{
library(sf)

# Example using calibration data
temp_path <- system.file("extdata", "calibration_data.xlsx", package = "SWATprepR")
reach_path <- system.file("extdata", "GIS/reaches.shp", package = "SWATprepR")
basin_path <- system.file("extdata", "GIS/basin.shp", package = "SWATprepR")
cal_data <- load_template(temp_path, 4326)
reach <- st_transform(st_read(reach_path), 4326)
basin <- st_transform(st_read(basin_path), 4326)
plot_map(cal_data$data, cal_data$stations, reach, basin)
}
\keyword{plotting}
