% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_monthly}
\alias{plot_monthly}
\title{Aggregating values of single station to interactive monthly box plots}
\usage{
plot_monthly(df, station, variables = NULL)
}
\arguments{
\item{df}{Dataframe with formatted data (Station, DATE, Variables and Values columns are needed).
Data can be loaded with \code{\link{load_template}} function using 'xlsx' template file.}

\item{station}{Character, indicating station ID, which should be selected for figure.}

\item{variables}{(optional) Character vector, variables/s, which should be in figure.
Default is \code{variables = NULL}, which means all variables will be in figure.}
}
\value{
plotly object of interactive figure
}
\description{
Aggregating values of single station to interactive monthly box plots
}
\examples{
\dontrun{
temp_path <- system.file("extdata", "calibration_data.xlsx", package = "SWATprepR")
cal_data <- load_template(temp_path)
plot_monthly(cal_data$data, station = "4")
}
}
\keyword{plotting}
