% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_monthly}
\alias{plot_monthly}
\title{Aggregating values of single station to interactive monthly box plots}
\usage{
plot_monthly(df, station, variables = levels(as.factor(df$Variables)))
}
\arguments{
\item{df}{dataframe with formatted data (Station, DATE, Variables and Values columns are needed)}

\item{station}{character indicating station, which should be selected for figure.}

\item{variables}{optional parameter of character vector, which parameters should be in figure.}
}
\value{
plotly object of interactive figure
}
\description{
Aggregating values of single station to interactive monthly box plots
}
\examples{
temp_path <- system.file("extdata", "calibration_data.xlsx", package = "SWATprepR")
cal_data <- load_template(temp_path)
plot_monthly(cal_data$data, station = "4")
}
\keyword{plotting}
