% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_weather_compare}
\alias{plot_weather_compare}
\title{Plotting figure to comparing two datasets with weather data}
\usage{
plot_weather_compare(
  meteo_lst1,
  meteo_lst2,
  par,
  period = "day",
  fn_summarize = "mean",
  name_set1 = "dataset 1",
  name_set2 = "dataset 2"
)
}
\arguments{
\item{meteo_lst1}{first nested list of lists with dataframes.
Nested structure meteo_lst -> data -> Station ID -> Parameter -> Dataframe (DATE, PARAMETER).}

\item{meteo_lst2}{second nested list of lists with dataframes.
Nested structure meteo_lst -> data -> Station ID -> Parameter -> Dataframe (DATE, PARAMETER).}

\item{par}{character marking weather variable to extract (i.e. "PCP", "SLR", etc).}

\item{period}{character describing, which time interval to display (default is "day",
other examples are "week", "month", etc).}

\item{fn_summarize}{function to recalculate to time interval (default is "mean", other examples
are "median", "sum", etc).}

\item{name_set1}{character to name first dataset.}

\item{name_set2}{character to name second dataset.}
}
\value{
plotly figure object with displayed weather data for two datasets.
}
\description{
Plotting figure to comparing two datasets with weather data
}
\examples{
temp_path <- system.file("extdata", "weather_data.xlsx", package = "svatools")
met_lst1 <- load_template(temp_path, 4326)
temp_path <- system.file("extdata", "weather_data_raw.xlsx", package = "svatools")
met_lst2 <- load_template(temp_path, 4326)
plot_weather_compare(met_lst1, met_lst2, "PCP", "month", "mean", "clean", "raw")
}
