% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparing.R
\name{prepare_climate}
\alias{prepare_climate}
\title{Prepare or Update Climate Data Text Input Files in SWAT+ Model}
\usage{
prepare_climate(meteo_lst, write_path, period_starts = NA, period_ends = NA)
}
\arguments{
\item{meteo_lst}{Nested list with dataframes.
Nested structure: \code{meteo_lst -> data -> Station ID -> Parameter -> 
  Dataframe (DATE, PARAMETER)},
\code{meteo_lst -> stations -> Dataframe (ID, Name, Elevation, Source, 
  geometry, Long, Lat)}. \cr\cr
meteo_lst can be created using \code{\link{load_template}} function using
'xlsx' template file or it could to be created with \code{\link{load_swat_weather}}
function loading information from SWAT+ model setup weather files.}

\item{write_path}{Character, path to the SWAT+ txtinout folder (example "my_model").}

\item{period_starts}{(optional) Character, date string (example '1991-01-01').
Default \code{period_starts = NA}, stands for all available in data.}

\item{period_ends}{(optional) Character, date string (example '2020-12-31').
Default \code{period_ends = NA}, stands for all available in data.}
}
\value{
Fills or updates multiple weather-related text files in the SWAT+ model.
}
\description{
This function prepares or updates climate data text input files in a SWAT+
model based on the provided meteo_lst.
}
\examples{
\dontrun{
  # Example usage:
  prepare_climate(meteo_lst, "output", "1991-01-01", "2020-12-31")
}
}
\seealso{
\code{\link{load_template}}
}
\keyword{writing}
