% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparing.R
\name{usersoil_to_sol}
\alias{usersoil_to_sol}
\title{Convert usersoil.csv to soil.sol}
\usage{
usersoil_to_sol(csv_path, db_path = NULL)
}
\arguments{
\item{csv_path}{character path to csv file (example "usersoil_lrew.csv"")}

\item{db_path}{character to sqlite project database (example "output/project.sqlite").
Optional, default NULL. Could be used to reduce soils.sol file, if there are less soil
types in sqlite database than in usersoil csv file.}
}
\value{
soils.sol SWAT+ model input file
}
\description{
Convert usersoil.csv to soil.sol
}
\examples{
\dontrun{
usersoil_to_sol("output/usersoil_lrew.csv")
##Or
usersoil_to_sol("output/usersoil_lrew.csv", "output/project.sqlite")
}
}
